Issue #168: Fix issues with FTPS uploads to backend servers.
The problem was that we need to do network shutdowns on both read/write
streams first, and _then_ close those streams.  We were doing a shutdown+close
on the read stream, then on the write stream.  While this works for
downloads/data transfers, with regard to proper TLS session shutdowns, it is
the wrong order of events for uploads.