Fix compilation with OpenSSL prior to 1.1.x
Variable crypto_engine is used but not declared when build with
OpenSSL later than 0.9.7 and prior to 1.1.0. It's also necessary
to include <openssl/engine.h> for the declaration of ENGINE_cleanup()
in the same cases.

The related code looks like it might be left over and not actually
used in mod_proxy.